Restructure code for submission to PyPI. Change name of System class to QSystem.
