+++ README.md
@@ -1,2 +1,3 @@
 # magpy
-Magnus based time-integration for spins and qubits under a magnetic field
+Magnus based time-integration for spins and qubits under a magnetic field. Please see the example Jupyter notebook for a walkthrough of 
+Magpy's functionality.
