+++ magpy/__init__.py
@@ -1,0 +1,3 @@
+from .methods import *
+from .HOp import HOp
+from .QSystem import QSystem
