+++ pyproject.toml
@@ -1,0 +1,27 @@
+[build-system]
+requires = ["numpy", "scipy"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "magpy"
+authors = [
+    { name = "Danny Goodacre", email = "dannygoodacre1@gmail.com"},
+    { name = "Pranav Singh", email = "ps2106@bath.ac.uk"}
+]
+description = "Magnus based time-integration for spins and qubits under a magnetic field"
+readme = "README.md"
+license = { file="LICENSE"}
+requires-python = ">=3.8.10"
+requires = [
+    "numpy>=1.23.2",
+    "scipy>=1.9.0"
+]
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+]
+
+[project.urls]
+"Homepage" = "https://github.com/brownadder/magpy"
+"Bug Tracker" = "https://github.com/brownadder/magpy/issues"
